feat: currentGameType 에 맞는 handler 추가
